feat: adds new sub-component
