+++ frontend/public/locales/sv/common.json
@@ -21,6 +21,7 @@
   "done": "Klart",
   "cancel": "Avbryt",
   "delete": "Radera",
+  "help": "Hjälp",
   "yesContinue": "Ja, fortsätt",
   "dispatchDone": "Ditt utskick har gjorts",
   "dispatchNew": "Gör ett nytt utskick",
